+++ click_extended/__init__.py
@@ -1,7 +1,9 @@
 """Initialization file for the 'click_extended' module."""
 
 from click_extended.core._child_node import ChildNode
+from click_extended.core._global_node import GlobalNode
 from click_extended.core._parent_node import ParentNode
+from click_extended.core._root_node import RootNode
 from click_extended.core.argument import argument
 from click_extended.core.command import command
 from click_extended.core.env import env
@@ -11,7 +13,9 @@ from click_extended.core.tag import tag
 
 __all__ = [
     "ChildNode",
+    "GlobalNode",
     "ParentNode",
+    "RootNode",
     "argument",
     "command",
     "env",